# resources/openapi/georchestra-plui-evolution.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Georchestra - Plugin Plui'Evolution' API",
        "contact": {
            "email": "thomas.brule@open-groupe.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
	"servers": [
		{
			"url": "/pluievolution",
			"description": "Georchestra - Plugin PluiEvolution API"
		}
	],    
...
    "paths": {
...
        "/carto/wfsRequest": {
            "get": {
                "tags": [
                    "carto"
                ],
                "summary": "Requete WFS get (essentiellement pour la description des features)",
                "description": "Requete WFS de récupération des informations sur le service",
                "operationId": "getWfs",
                "responses": {
                    "200": {
                        "description": "Reponse du geoserver",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "text/html": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "text/html": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "carto"
                ],
                "summary": "Requete WFS post (notamment pour les features)",
                "description": "Requete WFS de recuperation d'un claque",
                "operationId": "postWfs",
                "requestBody": {
                    "description": "Contenu WFS au format XML",
                    "content": {
                        "*/*": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "text/html": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "text/html": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "application/octet-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            },
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "x-codegen-request-body-name": "wfsContent"
            }
        },
        "/carto/layerConfiguration": {
            "get": {
                "tags": [
                    "carto"
                ],
                "summary": "Configurations du layer geoserver",
                "operationId": "getLayerConfigurations",
                "responses": {
                    "200": {
                        "description": "Configuration du layer",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/LayerConfiguration"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    }
                }
            },
...
            "LayerConfiguration": {
                "type": "object",
                "properties": {
                    "layerWorkspace": {
                        "type": "string"
                    },
                    "layerName": {
                        "type": "string"
                    },
                    "layerProjection": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}